TV: Move PlaylistsViewModel to its own file
